net_cliente

+++ Models/Clientes.cs
@@ -2,7 +2,7 @@ using System.ComponentModel.DataAnnotations;
 namespace ApiEmpresa.Models;
 public class Clientes{
 [Key]
-public long Id { get; set; }
+public Int32 Id { get; set; }
 public string? nombres{get;set;}
 public string? apellidos{get;set;}
 public string? direccion{get;set;}
